+++ src/controllers/ElementsController.php
@@ -2862,7 +2862,10 @@ JS, [
             return false;
         }
 
-        if ($element->isProvisionalDraft) {
+        if (
+            $element->isProvisionalDraft &&
+            (!$element instanceof NestedElementInterface || !$element->getOwnerId())
+        ) {
             $element = $element->getCanonical(true);
         }